# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PA School Data API",
    "version": "1.0.0",
    "description": "Read-only JSON API behind chrispeterkins.com/paschools. Pennsylvania PSSA and Keystone results (2013/2015 to present), PVAAS growth, school directory data, and student-weighted aggregates. Responses are cacheable for an hour and carry ETags. Data: Pennsylvania Department of Education, NCES Common Core of Data, U.S. Census Bureau. See /paschools/about for methods and caveats."
  },
  "servers": [
    {
      "url": "https://chrispeterkins.com/paschools/api"
    }
  ],
...
  "components": {
    "parameters": {
      "exam": {
        "name": "exam",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "pssa",
            "keystone"
          ],
          "default": "pssa"
        },
        "description": "PSSA (grades 3-8) or Keystone (grade 11)"
      },
...
    },
...
  },
  "paths": {
...
    "/indicators/school/{id}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "Future Ready PA Index indicators, graduation rate, and enrollment history for one school",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "indicators: series per indicator with statewide value and a percentile among schools; enrollment by year; groups: latest attendance and graduation by student group with the gap vs. All Students"
          }
        }
      }
    },
    "/indicators/district/{id}": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "District graduation rate, enrollment-weighted school indicators, enrollment, and AFR spending per pupil by year",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "indicators (with percentiles), enrollment, finance per year vs the state, staff (teachers, students per teacher, average salary and experience vs the state), groups"
          }
        }
      }
    },
...
    "/indicators/spending": {
      "get": {
        "tags": [
          "indicators"
        ],
        "summary": "Every district's spending per pupil next to its all-grades Math + ELA proficiency, for a scatter",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Fiscal year end (2025 = 2024-25); defaults to the newest with both finance and results"
          },
          {
            "$ref": "#/components/parameters/exam"
          }
        ],
        "responses": {
          "200": {
            "description": "year, years, districts[{id,name,county,perPupil,instructionPerPupil,adm,proficiency,tested}], state{proficiency,medianPerPupil}"
          }
        }
      }
    },
...
  }
}